[scheduler][joy] Move the Popover.Trigger outside of the event components
