+++ packages/x-scheduler/src/joy/internals/utils/EventPopover.tsx
@@ -1,0 +1,101 @@
+'use client';
+import * as React from 'react';
+import { Popover } from '@base-ui-components/react/popover';
+import { CalendarEvent } from '../../models/events';
+import { EventPopover } from '../../event-popover/EventPopover';
+import { selectors } from '../../event-calendar/store';
+import { useEventCalendarStore } from '../hooks/useEventCalendarStore';
+import { useSelector } from '../../../base-ui-copy/utils/store';
+import { useEventCallback } from '../../../base-ui-copy/utils/useEventCallback';
+
+interface EventPopoverContext {
+  startEditing: (event: React.MouseEvent, calendarEvent: CalendarEvent) => void;
+}
+
+type EventPopoverProviderProps = {
+  containerRef: React.RefObject<HTMLElement | null>;
+  children: React.ReactNode;
+  onEventsChange?: (value: CalendarEvent[]) => void;
+};
+
+const EventPopoverContext = React.createContext<EventPopoverContext>({
+  startEditing: () => {},
+});
+
+export function EventPopoverProvider({
+  containerRef,
+  onEventsChange,
+  children,
+}: EventPopoverProviderProps) {
+  const [isPopoverOpen, setIsPopoverOpen] = React.useState(false);
+  const [anchor, setAnchor] = React.useState<HTMLElement | null>(null);
+  const [selectedEvent, setSelectedEvent] = React.useState<CalendarEvent | null>(null);
+  const store = useEventCalendarStore();
+  const resourcesByIdMap = useSelector(store, selectors.resourcesByIdMap);
+
+  const startEventEditing = useEventCallback(
+    (event: React.MouseEvent, calendarEvent: CalendarEvent) => {
+      setAnchor(event.currentTarget as HTMLElement);
+      setSelectedEvent(calendarEvent);
+      setIsPopoverOpen(true);
+    },
+  );
+
+  const handleClose = useEventCallback(() => {
+    if (!isPopoverOpen) {
+      return;
+    }
+    setIsPopoverOpen(false);
+    setAnchor(null);
+    setSelectedEvent(null);
+  });
+
+  const handleEventEdit = useEventCallback((editedEvent: CalendarEvent) => {
+    const prevEvents = store.state.events;
+    const updatedEvents = prevEvents.map((ev) => (ev.id === editedEvent.id ? editedEvent : ev));
+
+    if (onEventsChange) {
+      onEventsChange(updatedEvents);
+    }
+  });
+
+  const handleEventDelete = useEventCallback((deletedEventId: string) => {
+    const prevEvents = store.state.events;
+    const updatedEvents = prevEvents.filter((ev) => ev.id !== deletedEventId);
+
+    if (onEventsChange) {
+      onEventsChange(updatedEvents);
+    }
+  });
+
+  const contextValue = React.useMemo<EventPopoverContext>(
+    () => ({ startEditing: startEventEditing }),
+    [],
+  );
+
+  return (
+    <EventPopoverContext.Provider value={contextValue}>
+      <Popover.Root open={isPopoverOpen} onOpenChange={handleClose} modal>
+        {children}
+        {anchor && selectedEvent && (
+          <EventPopover
+            anchor={anchor}
+            calendarEvent={selectedEvent}
+            calendarEventResource={resourcesByIdMap.get(selectedEvent.resource)}
+            container={containerRef.current}
+            onEventEdit={handleEventEdit}
+            onEventDelete={handleEventDelete}
+            onClose={handleClose}
+          />
+        )}
+      </Popover.Root>
+    </EventPopoverContext.Provider>
+  );
+}
+
+export function EventPopoverTrigger(props: Popover.Trigger.Props & { event: CalendarEvent }) {
+  const { event: calendarEvent, ...other } = props;
+  const { startEditing } = React.useContext(EventPopoverContext);
+
+  return <Popover.Trigger onClick={(event) => startEditing(event, calendarEvent)} {...other} />;
+}
