+++ packages/x-scheduler/src/joy/month-view/MonthView.tsx
@@ -12,7 +12,7 @@ import { selectors } from '../event-calendar/store';
 import { useWeekList } from '../../primitives/use-week-list/useWeekList';
 import { DayGrid } from '../../primitives/day-grid';
 import { DayGridEvent } from '../internals/components/event/day-grid-event/DayGridEvent';
-import { EventPopoverProvider } from '../internals/utils/EventPopoverProvider';
+import { EventPopoverProvider, EventPopoverTrigger } from '../internals/utils/EventPopover';
 import { SchedulerValidDate } from '../../primitives/models';
 import { isWeekend } from '../internals/utils/date-utils';
 import { useTranslations } from '../internals/utils/TranslationsContext';
@@ -93,87 +93,89 @@ export const MonthView = React.memo(
     return (
       <div ref={handleRef} className={clsx('MonthViewContainer', 'joy', className)} {...other}>
         <EventPopoverProvider containerRef={containerRef} onEventsChange={onEventsChange}>
-          {({ onEventClick }) => (
-            <DayGrid.Root className="MonthViewRoot">
-              <div className="MonthViewHeader">
-                <div className="MonthViewWeekHeaderCell">{translations.weekAbbreviation}</div>
-                {weeks[0].map((day) => (
-                  <div
-                    key={day.date.toString()}
-                    id={`MonthViewHeaderCell-${day.date.toString()}`}
-                    role="columnheader"
-                    className="MonthViewHeaderCell"
-                    aria-label={adapter.format(day.date, 'weekday')}
-                  >
-                    {adapter.formatByString(day.date, 'ccc')}
-                  </div>
-                ))}
-              </div>
-              <div className="MonthViewBody">
-                {weeks.map((week, weekIdx) => {
-                  const weekNumer = adapter.getWeekNumber(week[0].date);
-                  return (
-                    <DayGrid.Row key={weekNumer} className="MonthViewRow">
-                      <div
-                        className="MonthViewWeekNumberCell"
-                        role="rowheader"
-                        aria-label={translations.weekNumberAriaLabel(weekNumer)}
-                      >
-                        {weekNumer}
-                      </div>
-                      {week.map((day, dayIdx) => {
-                        const isCurrentMonth = adapter.isSameMonth(day.date, visibleDate);
-                        const isToday = adapter.isSameDay(day.date, today);
+          <DayGrid.Root className="MonthViewRoot">
+            <div className="MonthViewHeader">
+              <div className="MonthViewWeekHeaderCell">{translations.weekAbbreviation}</div>
+              {weeks[0].map((day) => (
+                <div
+                  key={day.date.toString()}
+                  id={`MonthViewHeaderCell-${day.date.toString()}`}
+                  role="columnheader"
+                  className="MonthViewHeaderCell"
+                  aria-label={adapter.format(day.date, 'weekday')}
+                >
+                  {adapter.formatByString(day.date, 'ccc')}
+                </div>
+              ))}
+            </div>
+            <div className="MonthViewBody">
+              {weeks.map((week, weekIdx) => {
+                const weekNumer = adapter.getWeekNumber(week[0].date);
+                return (
+                  <DayGrid.Row key={weekNumer} className="MonthViewRow">
+                    <div
+                      className="MonthViewWeekNumberCell"
+                      role="rowheader"
+                      aria-label={translations.weekNumberAriaLabel(weekNumer)}
+                    >
+                      {weekNumer}
+                    </div>
+                    {week.map((day, dayIdx) => {
+                      const isCurrentMonth = adapter.isSameMonth(day.date, visibleDate);
+                      const isToday = adapter.isSameDay(day.date, today);
 
-                        const visibleEvents = day.events.slice(0, maxEvents);
-                        const hiddenCount = day.events.length - maxEvents;
-                        return (
-                          <DayGrid.Cell
-                            ref={weekIdx === 0 && dayIdx === 0 ? cellRef : undefined}
-                            key={day.date.toString()}
-                            className={clsx(
-                              'MonthViewCell',
-                              !isCurrentMonth && 'OtherMonth',
-                              isToday && 'Today',
-                              isWeekend(adapter, day.date) && 'Weekend',
-                            )}
-                          >
-                            {onDayHeaderClick ? (
-                              <button
-                                type="button"
-                                className="MonthViewCellNumberButton"
-                                onClick={handleHeaderClick(day.date)}
-                                tabIndex={0}
-                              >
-                                {renderCellNumberContent(day.date)}
-                              </button>
-                            ) : (
-                              renderCellNumberContent(day.date)
-                            )}
-                            {visibleEvents.map((eventProp) => (
-                              <DayGridEvent
-                                key={eventProp.id}
-                                event={eventProp}
-                                eventResource={resourcesByIdMap.get(eventProp.resource)}
-                                variant="compact"
-                                ariaLabelledBy={`MonthViewHeaderCell-${day.date.toString()}`}
-                                onEventClick={onEventClick}
-                              />
-                            ))}
-                            {hiddenCount > 0 && day.events.length > 0 && (
-                              <p className="MonthViewMoreEvents">
-                                {translations.hiddenEvents(hiddenCount)}
-                              </p>
-                            )}
-                          </DayGrid.Cell>
-                        );
-                      })}
-                    </DayGrid.Row>
-                  );
-                })}
-              </div>
-            </DayGrid.Root>
-          )}
+                      const visibleEvents = day.events.slice(0, maxEvents);
+                      const hiddenCount = day.events.length - maxEvents;
+                      return (
+                        <DayGrid.Cell
+                          ref={weekIdx === 0 && dayIdx === 0 ? cellRef : undefined}
+                          key={day.date.toString()}
+                          className={clsx(
+                            'MonthViewCell',
+                            !isCurrentMonth && 'OtherMonth',
+                            isToday && 'Today',
+                            isWeekend(adapter, day.date) && 'Weekend',
+                          )}
+                        >
+                          {onDayHeaderClick ? (
+                            <button
+                              type="button"
+                              className="MonthViewCellNumberButton"
+                              onClick={handleHeaderClick(day.date)}
+                              tabIndex={0}
+                            >
+                              {renderCellNumberContent(day.date)}
+                            </button>
+                          ) : (
+                            renderCellNumberContent(day.date)
+                          )}
+                          {visibleEvents.map((event) => (
+                            <EventPopoverTrigger
+                              key={event.id}
+                              event={event}
+                              render={
+                                <DayGridEvent
+                                  event={event}
+                                  eventResource={resourcesByIdMap.get(event.resource)}
+                                  variant="compact"
+                                  ariaLabelledBy={`MonthViewHeaderCell-${day.date.toString()}`}
+                                />
+                              }
+                            />
+                          ))}
+                          {hiddenCount > 0 && day.events.length > 0 && (
+                            <p className="MonthViewMoreEvents">
+                              {translations.hiddenEvents(hiddenCount)}
+                            </p>
+                          )}
+                        </DayGrid.Cell>
+                      );
+                    })}
+                  </DayGrid.Row>
+                );
+              })}
+            </div>
+          </DayGrid.Root>
         </EventPopoverProvider>
       </div>
     );
